fix(datepicker): remove component export toggle
