+++ packages/react/src/components/DatePicker/DatePicker.js
@@ -6,8 +6,8 @@
  */
 
 import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import classNames from 'classnames';
+import React, { useEffect, useRef } from 'react';
+import cx from 'classnames';
 import flatpickr from 'flatpickr';
 import l10n from 'flatpickr/dist/l10n/index';
 import DatePickerInput from '../DatePickerInput';
@@ -15,11 +15,11 @@ import carbonFlatpickrAppendToPlugin from './plugins/appendToPlugin';
 import carbonFlatpickrFixEventsPlugin from './plugins/fixEventsPlugin';
 import carbonFlatpickrRangePlugin from './plugins/rangePlugin';
 import { match, keys } from '../../internal/keyboard';
-import { FeatureFlagContext } from '../FeatureFlags';
-import { PrefixContext } from '../../internal/usePrefix';
+import { usePrefix } from '../../internal/usePrefix';
+import { useSavedCallback } from '../../internal/useSavedCallback';
 
 // Weekdays shorthand for english locale
-l10n.en.weekdays.shorthand.forEach((day, index) => {
+l10n.en.weekdays.shorthand.forEach((_day, index) => {
   const currentDay = l10n.en.weekdays.shorthand;
   if (currentDay[index] === 'Thu' || currentDay[index] === 'Th') {
     currentDay[index] = 'Th';
@@ -115,578 +115,529 @@ const carbonFlatpickrMonthSelectPlugin = (config) => (fp) => {
   };
 };
 
-export default class DatePicker extends Component {
-  static propTypes = {
-    /**
-     * flatpickr prop passthrough. Allows the user to enter a date directly
-     * into the input field
-     */
-    allowInput: PropTypes.bool,
+/**
+ * Determine if every child in a list of children has no label specified
+ * @param {Array<ReactElement>} children
+ * @returns {boolean}
+ */
+function isLabelTextEmpty(children) {
+  return children.every((child) => !child.props.labelText);
+}
 
-    /**
-     * The DOM element the Flatpicker should be inserted into. `<body>` by default.
-     */
-    appendTo: PropTypes.object,
+const rightArrowHTML = `<svg width="16px" height="16px" viewBox="0 0 16 16">
+  <polygon points="11,8 6,13 5.3,12.3 9.6,8 5.3,3.7 6,3 "/>
+  <rect width="16" height="16" style="fill:none" />
+</svg>`;
 
-    /**
-     * The child nodes.
-     */
-    children: PropTypes.node,
+const leftArrowHTML = `<svg width="16px" height="16px" viewBox="0 0 16 16">
+  <polygon points="5,8 10,3 10.7,3.7 6.4,8 10.7,12.3 10,13 "/>
+  <rect width="16" height="16" style="fill:none" />
+</svg>`;
 
-    /**
-     * The CSS class names.
-     */
-    className: PropTypes.string,
-
-    /**
-     * The date format.
-     */
-    dateFormat: PropTypes.string,
-
-    /**
-     * The type of the date picker:
-     *
-     * * `simple` - Without calendar dropdown.
-     * * `single` - With calendar dropdown and single date.
-     * * `range` - With calendar dropdown and a date range.
-     */
-    datePickerType: PropTypes.oneOf(['simple', 'single', 'range']),
-
-    /**
-     * The flatpickr `disable` option that allows a user to disable certain dates.
-     */
-    disable: PropTypes.array,
-
-    /**
-     * The flatpickr `enable` option that allows a user to enable certain dates.
-     */
-    enable: PropTypes.array,
-
-    /**
-     * `true` to use the light version.
-     */
-    light: PropTypes.bool,
-
-    /**
-     *  The language locale used to format the days of the week, months, and numbers. The full list of supported locales can be found here https://github.com/flatpickr/flatpickr/tree/master/src/l10n
-     */
-    locale: PropTypes.oneOfType([
-      PropTypes.object,
-      PropTypes.oneOf([
-        'ar', // Arabic
-        'at', // Austria
-        'az', // Azerbaijan
-        'be', // Belarusian
-        'bg', // Bulgarian
-        'bn', // Bangla
-        'bs', // Bosnia
-        'cat', // Catalan
-        'cs', // Czech
-        'cy', // Welsh
-        'da', // Danish
-        'de', // German
-        'en', // English
-        'eo', // Esperanto
-        'es', // Spanish
-        'et', // Estonian
-        'fa', // Persian
-        'fi', // Finnish
-        'fo', // Faroese
-        'fr', // French
-        'ga', // Gaelic
-        'gr', // Greek
-        'he', // Hebrew
-        'hi', // Hindi
-        'hr', // Croatian
-        'hu', // Hungarian
-        'id', // Indonesian
-        'is', // Icelandic
-        'it', // Italian
-        'ja', // Japanese
-        'ka', // Georgian
-        'km', // Khmer
-        'ko', // Korean
-        'kz', // Kazakh
-        'lt', // Lithuanian
-        'lv', // Latvian
-        'mk', // Macedonian
-        'mn', // Mongolian
-        'ms', // Malaysian
-        'my', // Burmese
-        'nl', // Dutch
-        'no', // Norwegian
-        'pa', // Punjabi
-        'pl', // Polish
-        'pt', // Portuguese
-        'ro', // Romanian
-        'ru', // Russian
-        'si', // Sinhala
-        'sk', // Slovak
-        'sl', // Slovenian
-        'sq', // Albanian
-        'sr', // Serbian
-        'sv', // Swedish
-        'th', // Thai
-        'tr', // Turkish
-        'uk', // Ukrainian
-        'uz', // Uzbek
-        'uz_latn', // Uzbek Latin
-        'vn', // Vietnamese
-        'zh_tw', // Mandarin Traditional
-        'zh', // Mandarin
-      ]),
-    ]),
-
-    /**
-     * The maximum date that a user can pick to.
-     */
-    maxDate: PropTypes.string,
-
-    /**
-     * The minimum date that a user can start picking from.
-     */
-    minDate: PropTypes.string,
-
-    /**
-     * The `change` event handler.
-     */
-    onChange: PropTypes.func,
-
-    /**
-     * The `close` event handler.
-     */
-    onClose: PropTypes.func,
-
-    /**
-     * The `open` event handler.
-     */
-    onOpen: PropTypes.func,
-
-    /**
-     * `true` to use the short version.
-     */
-    short: PropTypes.bool,
-
-    /**
-     * The value of the date value provided to flatpickr, could
-     * be a date, a date number, a date string, an array of dates.
-     */
-    value: PropTypes.oneOfType([
-      PropTypes.string,
-      PropTypes.arrayOf(
-        PropTypes.oneOfType([
-          PropTypes.string,
-          PropTypes.number,
-          PropTypes.object,
-        ])
-      ),
-      PropTypes.object,
-      PropTypes.number,
-    ]),
-  };
-
-  static defaultProps = {
-    short: false,
-    light: false,
-    dateFormat: 'm/d/Y',
-    locale: 'en',
-  };
-
-  static contextType = PrefixContext;
-
-  componentDidMount() {
-    const {
-      allowInput,
-      appendTo,
-      datePickerType,
-      dateFormat,
-      locale,
-      minDate,
-      maxDate,
-      value,
-      onClose,
-      disable,
-      enable,
-      ...rest
-    } = this.props;
-    if (datePickerType === 'single' || datePickerType === 'range') {
-      const onHook = (electedDates, dateStr, instance) => {
-        this.updateClassNames(instance);
-      };
-
-      // Logic to determine if `enable` or `disable` will be passed down. If neither
-      // is provided, we return the default empty disabled array, allowing all dates.
-      let enableOrDisable = enable ? 'enable' : 'disable';
-      let enableOrDisableArr;
-      if (!enable && !disable) {
-        enableOrDisableArr = [];
-      } else if (enable) {
-        enableOrDisableArr = enable;
-      } else {
-        enableOrDisableArr = disable;
+function updateClassNames(calendar, prefix) {
+  const calendarContainer = calendar.calendarContainer;
+  const daysContainer = calendar.days;
+  if (calendarContainer && daysContainer) {
+    // calendarContainer and daysContainer are undefined if flatpickr detects a mobile device
+    calendarContainer.classList.add(`${prefix}--date-picker__calendar`);
+    calendarContainer
+      .querySelector('.flatpickr-month')
+      .classList.add(`${prefix}--date-picker__month`);
+    calendarContainer
+      .querySelector('.flatpickr-weekdays')
+      .classList.add(`${prefix}--date-picker__weekdays`);
+    calendarContainer
+      .querySelector('.flatpickr-days')
+      .classList.add(`${prefix}--date-picker__days`);
+    forEach.call(
+      calendarContainer.querySelectorAll('.flatpickr-weekday'),
+      (item) => {
+        const currentItem = item;
+        currentItem.innerHTML = currentItem.innerHTML.replace(/\s+/g, '');
+        currentItem.classList.add(`${prefix}--date-picker__weekday`);
       }
-
-      let localeData;
-      if (typeof locale === 'object') {
-        let location = locale.locale ? locale.locale : 'en';
-        localeData = { ...l10n[location], ...locale };
-      } else {
-        localeData = l10n[locale];
+    );
+    forEach.call(daysContainer.querySelectorAll('.flatpickr-day'), (item) => {
+      item.classList.add(`${prefix}--date-picker__day`);
+      if (
+        item.classList.contains('today') &&
+        calendar.selectedDates.length > 0
+      ) {
+        item.classList.add('no-border');
+      } else if (
+        item.classList.contains('today') &&
+        calendar.selectedDates.length === 0
+      ) {
+        item.classList.remove('no-border');
       }
+    });
+  }
+}
 
-      // inputField ref might not be set in enzyme tests
-      if (this.inputField) {
-        this.cal = new flatpickr(this.inputField, {
-          inline: rest.inline ?? false,
-          disableMobile: true,
-          defaultDate: value,
-          mode: datePickerType,
-          allowInput: allowInput ?? true,
-          dateFormat: dateFormat,
-          locale: localeData,
-          [enableOrDisable]: enableOrDisableArr,
-          minDate: minDate,
-          maxDate: maxDate,
-          plugins: [
-            datePickerType === 'range'
-              ? new carbonFlatpickrRangePlugin({
-                  input: this.toInputField,
-                })
-              : () => {},
-            appendTo
-              ? carbonFlatpickrAppendToPlugin({
-                  appendTo,
-                })
-              : () => {},
-            carbonFlatpickrMonthSelectPlugin({
-              selectorFlatpickrMonthYearContainer: '.flatpickr-current-month',
-              selectorFlatpickrYearContainer: '.numInputWrapper',
-              selectorFlatpickrCurrentMonth: '.cur-month',
-              classFlatpickrCurrentMonth: 'cur-month',
-            }),
-            carbonFlatpickrFixEventsPlugin({
-              inputFrom: this.inputField,
-              inputTo: this.toInputField,
-            }),
-          ],
-          clickOpens: true,
-          nextArrow: this.rightArrowHTML(),
-          prevArrow: this.leftArrowHTML(),
-          onChange: (...args) => {
-            const { onChange } = this.props;
-            if (onChange) {
-              onChange(...args);
-            }
-          },
-          onClose,
-          onReady: onHook,
-          onMonthChange: onHook,
-          onYearChange: onHook,
-          onOpen: (...args) => {
-            const { onOpen } = this.props;
-            onHook(...args);
-            if (onOpen) {
-              onOpen(...args);
-            }
-          },
-          onValueUpdate: onHook,
+function DatePicker({
+  allowInput,
+  appendTo,
+  children,
+  className,
+  dateFormat = 'm/d/Y',
+  datePickerType,
+  disable,
+  enable,
+  inline,
+  light = false,
+  locale = 'en',
+  maxDate,
+  minDate,
+  onChange,
+  onClose,
+  onOpen,
+  short = false,
+  value,
+  ...rest
+}) {
+  const prefix = usePrefix();
+  const startInputField = useRef(null);
+  const endInputField = useRef(null);
+  const calendarRef = useRef(null);
+  const savedOnChange = useSavedCallback(onChange);
+  const savedOnClose = useSavedCallback(onClose);
+  const savedOnOpen = useSavedCallback(onOpen);
+
+  const datePickerClasses = cx(`${prefix}--date-picker`, {
+    [`${prefix}--date-picker--short`]: short,
+    [`${prefix}--date-picker--light`]: light,
+    [`${prefix}--date-picker--simple`]: datePickerType === 'simple',
+    [`${prefix}--date-picker--single`]: datePickerType === 'single',
+    [`${prefix}--date-picker--range`]: datePickerType === 'range',
+    [`${prefix}--date-picker--nolabel`]:
+      datePickerType === 'range' && isLabelTextEmpty(children),
+  });
+  const wrapperClasses = cx(`${prefix}--form-item`, { [className]: className });
+
+  const childrenWithProps = React.Children.toArray(children).map(
+    (child, index) => {
+      if (
+        index === 0 &&
+        child.type === React.createElement(DatePickerInput, child.props).type
+      ) {
+        return React.cloneElement(child, {
+          datePickerType,
+          ref: startInputField,
         });
-        this.addKeyboardEvents(this.cal);
-        this.addRoleAttributeToDialog();
+      }
+      if (
+        index === 1 &&
+        child.type === React.createElement(DatePickerInput, child.props).type
+      ) {
+        return React.cloneElement(child, {
+          datePickerType,
+          ref: endInputField,
+        });
+      }
+      if (index === 0) {
+        return React.cloneElement(child, {
+          ref: startInputField,
+        });
+      }
+      if (index === 1) {
+        return React.cloneElement(child, {
+          ref: endInputField,
+        });
       }
     }
-  }
+  );
 
-  componentDidUpdate({
-    dateFormat: prevDateFormat,
-    minDate: prevMinDate,
-    maxDate: prevMaxDate,
-    value: prevValue,
-    disable: prevDisable,
-    enable: prevEnable,
-    ...prevRest
-  }) {
-    const {
-      dateFormat,
-      minDate,
-      maxDate,
-      value,
-      disable,
-      enable,
-      ...rest
-    } = this.props;
-    if (this.cal) {
-      if (prevDateFormat !== dateFormat) {
-        this.cal.set({ dateFormat });
-      }
-      if (prevMinDate !== minDate) {
-        this.cal.set('minDate', minDate);
-      }
-      if (prevMaxDate !== maxDate) {
-        this.cal.set('maxDate', maxDate);
-      }
-      if (disable !== prevDisable) {
-        this.cal.set('disable', disable);
-      }
-      if (enable !== prevEnable) {
-        this.cal.set('enable', enable);
-      }
-      if (rest.inline && rest.inline !== prevRest?.inline) {
-        this.cal.set('inline', rest.inline);
-      }
+  useEffect(() => {
+    if (datePickerType !== 'single' && datePickerType !== 'range') {
+      return;
     }
 
-    // Coordinate when the given `value` prop changes. When this happens, we
-    // should update the calendar to the new value.
-    if (prevValue !== value) {
-      if (this.cal) {
-        this.cal.setDate(this.props.value);
-        this.updateClassNames(this.cal);
-      } else if (this.inputField) {
-        this.inputField.value = this.props.value;
+    if (startInputField.current === null) {
+      return;
+    }
+
+    const onHook = (_electedDates, _dateStr, instance, prefix) => {
+      updateClassNames(instance, prefix);
+    };
+
+    // Logic to determine if `enable` or `disable` will be passed down. If neither
+    // is provided, we return the default empty disabled array, allowing all dates.
+    let enableOrDisable = enable ? 'enable' : 'disable';
+    let enableOrDisableArr;
+    if (!enable && !disable) {
+      enableOrDisableArr = [];
+    } else if (enable) {
+      enableOrDisableArr = enable;
+    } else {
+      enableOrDisableArr = disable;
+    }
+
+    let localeData;
+    if (typeof locale === 'object') {
+      let location = locale.locale ? locale.locale : 'en';
+      localeData = { ...l10n[location], ...locale };
+    } else {
+      localeData = l10n[locale];
+    }
+
+    const { current: start } = startInputField;
+    const { current: end } = endInputField;
+    const calendar = new flatpickr(start, {
+      inline: inline ?? false,
+      disableMobile: true,
+      defaultDate: value,
+      mode: datePickerType,
+      allowInput: allowInput ?? true,
+      dateFormat: dateFormat,
+      locale: localeData,
+      [enableOrDisable]: enableOrDisableArr,
+      minDate: minDate,
+      maxDate: maxDate,
+      plugins: [
+        datePickerType === 'range'
+          ? new carbonFlatpickrRangePlugin({
+              input: endInputField.current,
+            })
+          : () => {},
+        appendTo
+          ? carbonFlatpickrAppendToPlugin({
+              appendTo,
+            })
+          : () => {},
+        carbonFlatpickrMonthSelectPlugin({
+          selectorFlatpickrMonthYearContainer: '.flatpickr-current-month',
+          selectorFlatpickrYearContainer: '.numInputWrapper',
+          selectorFlatpickrCurrentMonth: '.cur-month',
+          classFlatpickrCurrentMonth: 'cur-month',
+        }),
+        carbonFlatpickrFixEventsPlugin({
+          inputFrom: startInputField.current,
+          inputTo: endInputField.current,
+        }),
+      ],
+      clickOpens: true,
+      nextArrow: rightArrowHTML,
+      prevArrow: leftArrowHTML,
+      onChange: (...args) => {
+        if (savedOnChange) {
+          savedOnChange(...args);
+        }
+      },
+      onClose: savedOnClose,
+      onReady: onHook,
+      onMonthChange: onHook,
+      onYearChange: onHook,
+      onOpen: (...args) => {
+        onHook(...args);
+        savedOnOpen(...args);
+      },
+      onValueUpdate: onHook,
+    });
+
+    calendarRef.current = calendar;
+
+    function handleArrowDown(event) {
+      if (match(event, keys.ArrowDown)) {
+        const {
+          calendarContainer,
+          selectedDateElem: fpSelectedDateElem,
+          todayDateElem: fptodayDateElem,
+        } = calendar;
+        const selectedDateElem =
+          calendarContainer.querySelector('.selected') && fpSelectedDateElem;
+        const todayDateElem =
+          calendarContainer.querySelector('.today') && fptodayDateElem;
+        (
+          selectedDateElem ||
+          todayDateElem ||
+          calendarContainer.querySelector('.flatpickr-day[tabindex]') ||
+          calendarContainer
+        ).focus();
       }
     }
-  }
 
-  componentWillUnmount() {
-    if (this.cal) {
-      this.cal.destroy();
-    }
-    if (this.inputField) {
-      this.inputField.removeEventListener('change', this.onChange);
-    }
-    if (this.toInputField) {
-      this.toInputField.removeEventListener('change', this.onChange);
-    }
-  }
+    function handleOnChange() {
+      if (start.value !== '') {
+        return;
+      }
 
-  onChange = () => {
-    if (this.inputField.value === '' && this.cal?.selectedDates.length) {
-      this.cal.clear();
-      this.cal.input.focus();
-    }
-  };
+      if (!calendar.selectedDates) {
+        return;
+      }
 
-  /**
-   * Flatpickr's calendar dialog is not rendered in a landmark causing an
-   * error with IBM Equal Access Accessibility Checker so we add an aria
-   * role to the container div.
-   */
-  addRoleAttributeToDialog = () => {
-    if (this.inputField) {
-      this.cal.calendarContainer.setAttribute('role', 'region');
+      if (calendar.selectedDates.length === 0) {
+        return;
+      }
+
+      calendar.clear();
+      calendar.input.focus();
+    }
+
+    if (start) {
+      start.addEventListener('keydown', handleArrowDown);
+      start.addEventListener('change', handleOnChange);
+
+      // Flatpickr's calendar dialog is not rendered in a landmark causing an
+      // error with IBM Equal Access Accessibility Checker so we add an aria
+      // role to the container div.
+      calendar.calendarContainer.setAttribute('role', 'region');
       // IBM EAAC requires an aria-label on a role='region'
-      this.cal.calendarContainer.setAttribute(
+      calendar.calendarContainer.setAttribute(
         'aria-label',
         'calendar-container'
       );
     }
-  };
 
-  addKeyboardEvents = (cal) => {
-    const initArrowDownListener = (element) => {
-      if (element) {
-        element.addEventListener('keydown', (e) => {
-          if (match(e, keys.ArrowDown)) {
-            const {
-              calendarContainer,
-              selectedDateElem: fpSelectedDateElem,
-              todayDateElem: fptodayDateElem,
-            } = cal;
-            const selectedDateElem =
-              calendarContainer.querySelector('.selected') &&
-              fpSelectedDateElem;
-            const todayDateElem =
-              calendarContainer.querySelector('.today') && fptodayDateElem;
-            (
-              selectedDateElem ||
-              todayDateElem ||
-              calendarContainer.querySelector('.flatpickr-day[tabindex]') ||
-              calendarContainer
-            ).focus();
-          }
-        });
-        element.addEventListener('change', this.onChange);
+    if (end) {
+      end.addEventListener('keydown', handleArrowDown);
+      end.addEventListener('change', handleOnChange);
+    }
+
+    //component did unmount equivalent
+    return () => {
+      // Note: if the `startInputField` ref is undefined then calendar will be
+      // of type: Array and `destroy` will not be defined
+      if (calendar && calendar.destroy) {
+        calendar.destroy();
+      }
+
+      if (start) {
+        start.removeEventListener('keydown', handleArrowDown);
+        start.removeEventListener('change', handleOnChange);
+      }
+
+      if (end) {
+        end.removeEventListener('keydown', handleArrowDown);
+        end.removeEventListener('change', handleOnChange);
       }
     };
-    initArrowDownListener(this.inputField);
-    initArrowDownListener(this.toInputField);
-  };
+  }, [savedOnChange, savedOnClose, savedOnOpen]); //eslint-disable-line react-hooks/exhaustive-deps
 
-  rightArrowHTML() {
-    return `
-      <svg width="16px" height="16px" viewBox="0 0 16 16">
-        <polygon points="11,8 6,13 5.3,12.3 9.6,8 5.3,3.7 6,3 "/>
-        <rect width="16" height="16" style="fill:none" />
-      </svg>`;
-  }
-
-  leftArrowHTML() {
-    return `
-      <svg width="16px" height="16px" viewBox="0 0 16 16">
-        <polygon points="5,8 10,3 10.7,3.7 6.4,8 10.7,12.3 10,13 "/>
-        <rect width="16" height="16" style="fill:none" />
-      </svg>`;
-  }
-
-  updateClassNames = (calendar) => {
-    const calendarContainer = calendar.calendarContainer;
-    const daysContainer = calendar.days;
-    if (calendarContainer && daysContainer) {
-      // calendarContainer and daysContainer are undefined if flatpickr detects a mobile device
-      calendarContainer.classList.add(`${this.context}--date-picker__calendar`);
-      calendarContainer
-        .querySelector('.flatpickr-month')
-        .classList.add(`${this.context}--date-picker__month`);
-      calendarContainer
-        .querySelector('.flatpickr-weekdays')
-        .classList.add(`${this.context}--date-picker__weekdays`);
-      calendarContainer
-        .querySelector('.flatpickr-days')
-        .classList.add(`${this.context}--date-picker__days`);
-      forEach.call(
-        calendarContainer.querySelectorAll('.flatpickr-weekday'),
-        (item) => {
-          const currentItem = item;
-          currentItem.innerHTML = currentItem.innerHTML.replace(/\s+/g, '');
-          currentItem.classList.add(`${this.context}--date-picker__weekday`);
-        }
-      );
-      forEach.call(daysContainer.querySelectorAll('.flatpickr-day'), (item) => {
-        item.classList.add(`${this.context}--date-picker__day`);
-        if (
-          item.classList.contains('today') &&
-          calendar.selectedDates.length > 0
-        ) {
-          item.classList.add('no-border');
-        } else if (
-          item.classList.contains('today') &&
-          calendar.selectedDates.length === 0
-        ) {
-          item.classList.remove('no-border');
-        }
-      });
+  useEffect(() => {
+    if (calendarRef.current) {
+      calendarRef.current.set({ dateFormat });
     }
-  };
+  }, [dateFormat]);
 
-  assignInputFieldRef = (node) => {
-    this.inputField = !node
-      ? null
-      : // Child is a regular DOM node, seen in tests
-      node.nodeType === Node.ELEMENT_NODE
-      ? node.querySelector(`.${this.context}--date-picker__input`)
-      : // Child is a React component
-      node.input && node.input.nodeType === Node.ELEMENT_NODE
-      ? node.input
-      : null;
-  };
+  useEffect(() => {
+    if (calendarRef.current && minDate) {
+      calendarRef.current.set('minDate', minDate);
+    }
+  }, [minDate]);
 
-  assignToInputFieldRef = (node) => {
-    this.toInputField = !node
-      ? null
-      : // Child is a regular DOM node, seen in tests
-      node.nodeType === Node.ELEMENT_NODE
-      ? node.querySelector(`.${this.context}--date-picker__input`)
-      : // Child is a React component
-      node.input && node.input.nodeType === Node.ELEMENT_NODE
-      ? node.input
-      : null;
-  };
+  useEffect(() => {
+    if (calendarRef.current && maxDate) {
+      calendarRef.current.set('maxDate', maxDate);
+    }
+  }, [maxDate]);
 
-  isLabelTextEmpty = (children) =>
-    children.every((child) => !child.props.labelText);
+  useEffect(() => {
+    if (calendarRef.current && disable) {
+      calendarRef.current.set('disbale', disable);
+    }
+  }, [disable]);
 
-  render() {
-    return (
-      <FeatureFlagContext.Consumer>
-        {(scope) => {
-          const {
-            allowInput, // eslint-disable-line
-            appendTo, // eslint-disable-line
-            children,
-            className,
-            short,
-            light,
-            datePickerType,
-            minDate, // eslint-disable-line
-            maxDate, // eslint-disable-line
-            dateFormat, // eslint-disable-line
-            onChange, // eslint-disable-line
-            locale, // eslint-disable-line
-            value, // eslint-disable-line
-            onOpen, // eslint-disable-line
-            ...other
-          } = this.props;
+  useEffect(() => {
+    if (calendarRef.current && enable) {
+      calendarRef.current.set('enable', enable);
+    }
+  }, [enable]);
 
-          let enabled;
+  useEffect(() => {
+    if (calendarRef.current && inline) {
+      calendarRef.current.set('inline', inline);
+    }
+  }, [inline]);
 
-          if (scope.enabled) {
-            enabled = scope.enabled('enable-v11-release');
-          }
+  useEffect(() => {
+    if (calendarRef.current) {
+      calendarRef.current.set({ value });
+      updateClassNames(calendarRef.current, prefix);
+      //for simple date picker w/o calendar; initial mount may not have value
+    } else if (!calendarRef.current && value) {
+      startInputField.current.value = value;
+    }
+  }, [value, prefix]);
 
-          const datePickerClasses = classNames(
-            `${this.context}--date-picker`,
-            [enabled ? null : className],
-            {
-              [`${this.context}--date-picker--short`]: short,
-              [`${this.context}--date-picker--light`]: light,
-              [`${this.context}--date-picker--simple`]:
-                datePickerType === 'simple',
-              [`${this.context}--date-picker--single`]:
-                datePickerType === 'single',
-              [`${this.context}--date-picker--range`]:
-                datePickerType === 'range',
-              [`${this.context}--date-picker--nolabel`]:
-                datePickerType === 'range' && this.isLabelTextEmpty(children),
-            }
-          );
-
-          const wrapperClasses = classNames(`${this.context}--form-item`, [
-            enabled ? className : null,
-          ]);
-
-          const childArray = React.Children.toArray(children);
-          const childrenWithProps = childArray.map((child, index) => {
-            if (
-              index === 0 &&
-              child.type ===
-                React.createElement(DatePickerInput, child.props).type
-            ) {
-              return React.cloneElement(child, {
-                datePickerType,
-                ref: this.assignInputFieldRef,
-              });
-            }
-            if (
-              index === 1 &&
-              child.type ===
-                React.createElement(DatePickerInput, child.props).type
-            ) {
-              return React.cloneElement(child, {
-                datePickerType,
-                ref: this.assignToInputFieldRef,
-              });
-            }
-            if (index === 0) {
-              return React.cloneElement(child, {
-                ref: this.assignInputFieldRef,
-              });
-            }
-            if (index === 1) {
-              return React.cloneElement(child, {
-                ref: this.assignToInputFieldRef,
-              });
-            }
-          });
-          return (
-            <div className={wrapperClasses}>
-              <div className={datePickerClasses} {...other}>
-                {childrenWithProps}
-              </div>
-            </div>
-          );
-        }}
-      </FeatureFlagContext.Consumer>
-    );
-  }
+  return (
+    <div className={wrapperClasses} {...rest}>
+      <div className={datePickerClasses}>{childrenWithProps}</div>
+    </div>
+  );
 }
+
+DatePicker.propTypes = {
+  /**
+   * flatpickr prop passthrough. Allows the user to enter a date directly
+   * into the input field
+   */
+  allowInput: PropTypes.bool,
+
+  /**
+   * The DOM element the Flatpicker should be inserted into. `<body>` by default.
+   */
+  appendTo: PropTypes.object,
+
+  /**
+   * The child nodes.
+   */
+  children: PropTypes.node,
+
+  /**
+   * The CSS class names.
+   */
+  className: PropTypes.string,
+
+  /**
+   * The date format.
+   */
+  dateFormat: PropTypes.string,
+
+  /**
+   * The type of the date picker:
+   *
+   * * `simple` - Without calendar dropdown.
+   * * `single` - With calendar dropdown and single date.
+   * * `range` - With calendar dropdown and a date range.
+   */
+  datePickerType: PropTypes.oneOf(['simple', 'single', 'range']),
+
+  /**
+   * The flatpickr `disable` option that allows a user to disable certain dates.
+   */
+  disable: PropTypes.array,
+
+  /**
+   * The flatpickr `enable` option that allows a user to enable certain dates.
+   */
+  enable: PropTypes.array,
+
+  /**
+   * The flatpickr `inline` option.
+   */
+  inline: PropTypes.bool,
+
+  /**
+   * `true` to use the light version.
+   */
+  light: PropTypes.bool,
+
+  /**
+   *  The language locale used to format the days of the week, months, and numbers. The full list of supported locales can be found here https://github.com/flatpickr/flatpickr/tree/master/src/l10n
+   */
+  locale: PropTypes.oneOfType([
+    PropTypes.object,
+    PropTypes.oneOf([
+      'ar', // Arabic
+      'at', // Austria
+      'az', // Azerbaijan
+      'be', // Belarusian
+      'bg', // Bulgarian
+      'bn', // Bangla
+      'bs', // Bosnia
+      'cat', // Catalan
+      'cs', // Czech
+      'cy', // Welsh
+      'da', // Danish
+      'de', // German
+      'en', // English
+      'eo', // Esperanto
+      'es', // Spanish
+      'et', // Estonian
+      'fa', // Persian
+      'fi', // Finnish
+      'fo', // Faroese
+      'fr', // French
+      'ga', // Gaelic
+      'gr', // Greek
+      'he', // Hebrew
+      'hi', // Hindi
+      'hr', // Croatian
+      'hu', // Hungarian
+      'id', // Indonesian
+      'is', // Icelandic
+      'it', // Italian
+      'ja', // Japanese
+      'ka', // Georgian
+      'km', // Khmer
+      'ko', // Korean
+      'kz', // Kazakh
+      'lt', // Lithuanian
+      'lv', // Latvian
+      'mk', // Macedonian
+      'mn', // Mongolian
+      'ms', // Malaysian
+      'my', // Burmese
+      'nl', // Dutch
+      'no', // Norwegian
+      'pa', // Punjabi
+      'pl', // Polish
+      'pt', // Portuguese
+      'ro', // Romanian
+      'ru', // Russian
+      'si', // Sinhala
+      'sk', // Slovak
+      'sl', // Slovenian
+      'sq', // Albanian
+      'sr', // Serbian
+      'sv', // Swedish
+      'th', // Thai
+      'tr', // Turkish
+      'uk', // Ukrainian
+      'uz', // Uzbek
+      'uz_latn', // Uzbek Latin
+      'vn', // Vietnamese
+      'zh_tw', // Mandarin Traditional
+      'zh', // Mandarin
+    ]),
+  ]),
+
+  /**
+   * The maximum date that a user can pick to.
+   */
+  maxDate: PropTypes.string,
+
+  /**
+   * The minimum date that a user can start picking from.
+   */
+  minDate: PropTypes.string,
+
+  /**
+   * The `change` event handler.
+   */
+  onChange: PropTypes.func,
+
+  /**
+   * The `close` event handler.
+   */
+  onClose: PropTypes.func,
+
+  /**
+   * The `open` event handler.
+   */
+  onOpen: PropTypes.func,
+
+  /**
+   * `true` to use the short version.
+   */
+  short: PropTypes.bool,
+
+  /**
+   * The value of the date value provided to flatpickr, could
+   * be a date, a date number, a date string, an array of dates.
+   */
+  value: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.arrayOf(
+      PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number,
+        PropTypes.object,
+      ])
+    ),
+    PropTypes.object,
+    PropTypes.number,
+  ]),
+};
+
+export default DatePicker;
